+++ CHANGES.md
@@ -1,3 +1,5 @@
 
+  - improved `inclusion` (roughly 10 times faster)
+
 # 0.1.0
   - first release